+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.3.1 under development
 
-- no changes in this release.
+- Bug #768: Fix incorrect operation `Each (new Nested)` rule for iterable of objects (@Enjoyzz)
 
 ## 2.3.0 May 07, 2025